pyproject.toml
[project]
name = "agent-x"
version = "0.1.0"
readme="README.md"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "chromadb==1.5.5",
    "langchain==1.2.15",
    "langchain-community==0.4.1",
    "langchain-ollama==1.0.1",
    "langchain-openai==1.1.12",
    "langchain-pinecone==0.2.13",
    "langchain-tavily==0.2.17",
    "langchainhub==0.1.21",
    "pypdf==6.9.2",
    "python-dotenv==1.2.2",
    "langchain-experimental==0.4.1",
    "llama-cpp-python==0.3.20",
    "faiss-cpu==1.13.2",
    "langchain-chroma==1.1.0",
    "langgraph==1.1.6",
    "grandalf==0.8",
    "langchain-google-genai==4.2.1",
    "langchain-core==1.2.26",
    "langchain-openrouter==0.2.1",
    "openai==2.30.0",
    "langchain-mcp-adapters==0.2.2",
    "mcp[cli]==1.27.0",
    "aiohttp==3.13.5",
    "fastmcp==3.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests", "tests_sandbox"]
pythonpath = ["src"]
